@@ -51,6 +51,7 @@
   "author": "Boris Ding Poh Hing",
   "license": "MIT",
   "dependencies": {
+    "@std/esm": "^0.18.0",
     "app-root-path": "^2.0.1",
     "axios": "^0.17.1",
     "babel-plugin-react-css-modules": "^3.3.2",
@@ -141,9 +142,10 @@
     "stylelint-config-sass-guidelines": "^4.0.1",
     "stylelint-order": "^0.7.0",
     "stylelint-scss": "^2.1.0",
+    "uglifyjs-webpack-plugin": "^1.1.5",
     "url-loader": "^0.6.2",
-    "webpack": "^3.8.1",
-    "webpack-dev-middleware": "^1.12.2",
+    "webpack": "^3.10.0",
+    "webpack-dev-middleware": "^2.0.3",
     "webpack-hot-middleware": "^2.21.0",
     "webpack-hot-server-middleware": "^0.3.0"
   },
@@ -177,6 +179,10 @@
     "testPathIgnorePatterns": [
       "<rootDir>/node_modules/"
     ]
+  },
+  "@std/esm": {
+    "cjs": true,
+    "esm": "js"
   },
   "_moduleAliases": {
     "@root": ".",
