@@ -1,5 +1,8 @@
-// enable module aliases for api `require`
+// es module standard, which enables `import` for node
+require = require('@std/esm')(module);
+
+// enable module aliases for API `import` modules
 require('module-alias/register');
 
-// node server api entry
-require('./src/api');
+// entry for API
+module.exports = require('./src/api').default;
