@@ -1,20 +1,2 @@
-const chalk = require('chalk');
-
-module.exports = {
-  error: (message, code, substitution = []) => {
-    console.error(chalk.red(`${message}\n`), ...substitution);
-    code && process.exit(code);
-  },
-  info: (message, code, substitution = []) => {
-    console.info(chalk.cyanBright(`${message}\n`), ...substitution);
-    code && process.exit(code);
-  },
-  warn: (message, code, substitution = []) => {
-    console.warn(chalk.yellow(`${message}\n`), ...substitution);
-    code && process.exit(code);
-  },
-  success: (message, code, substitution = []) => {
-    console.log(chalk.green(`${message}\n`), ...substitution);
-    code && process.exit(code);
-  }
-};
+export { default as print } from './print';
+export { default as ServiceClass, service } from './service';
